Consolidate Activity & ActivityEx partial classes
